Fixes from initial commit
Mostly bug fixes, refactoring and changes to add the RNDR coin which was left out in the initial work. There are some files like the outlier-price-moves that come in two versions. These are used to generate sounds when large price moves occur. The versions are made to run on Python 3 and above and a legacy for Python < 3 as I needed it for one machine that I use.